Add arenas, skins, new weapons, and an in-ROM benchmark
Gameplay
- Three arenas (Construction Yard, Rocket Silo, Moon Temple), each unlocked
  by scoring 10 in the previous one, as in the original. The girder layer is
  the collision map, so an arena is a tilemap plus a row in the Arena table;
  spawns, crate spots and fire tiles are no longer hardcoded. The two new
  arenas reuse the existing tileset and differ by palette only, which works
  because the fire owns palette entries 1/8/9/14 exclusively.
- Eight unlockable skins, each with its own 14 animation tiles plus a
  four-entry palette swap. Tiles are generated by tools/genskins.py.
- Katana (weapon 9), using a sprite that was in the 2014 spritesheet but
  unreachable because the crate roll was rand() % 9. Turns large monsters
  around on hit, as in the original.
- Grenade launcher re-enabled; it needed bullet type 5 added to all three
  explosion sites or it was a lobbed dud.
- Explosions are a real radius (44px) drawn as an affine-scaled disc that
  grows, holds and collapses, replacing a spray of 30 shrapnel rounds whose
  damage depended on what happened to connect.
- Minigun fire rate 5 -> 2 frames, bullet pool 20 -> 28. Both sized by
  measurement; see .ai/experiments.md.
- High score, lifetime crates, equipped skin, chosen arena and per-arena best
  scores in cartridge SRAM, migrating forward from both older save formats.

Tooling
- tools/bench.sh runs a deterministic 30 second scripted benchmark in an
  emulator and reports avg/p95/max CPU, frame rate and what the run actually
  did. Results come out via SRAM, so it needs no UI interaction.
- tools/genskins.py and tools/genarenas.py generate playerskins.c and
  arenas.c; edit those rather than the output.
- tools/gba_sbrk.c is the heap and syscall support this build needs, since
  vanilla newlib has no _sbrk. See .ai/toolchain.md.

Fixes
- spawnBullet() returned slot 0 when the pool was exhausted, silently
  reversing a live bullet. Every caller audited; three fed the result
  straight into bullets.at(), which would have thrown.
- drawText() had no bounds check, and the credits page uses exactly its
  budget of objects.
- Bullet::setType() assigned its parameter to itself.
- Saving SRAM on every crate pickup caused a visible stutter; progress is now
  written only when it changes, and never during play.
- Object slots are derived from the pool sizes, so raising MAX_BULLETS no
  longer overwrites the text objects.

Background maps narrowed from int to unsigned short, saving 7.8KB of ROM.
Working notes in .ai/.

+++ background.h
@@ -28,16 +28,16 @@ extern const unsigned short backgroundPal[256];
 extern const unsigned short menuBGPal[256];
 
 #define girderMapLen 1024
-extern const int girderMap[1024];
+extern const unsigned short girderMap[1024];
 
 #define fenceBGMapLen 1024
-extern const int fenceBGMap[1024];
+extern const unsigned short fenceBGMap[1024];
 
 #define buildingBGMapLen 1024
-extern const int buildingBGMap[1024];
+extern const unsigned short buildingBGMap[1024];
 
 #define cloudBGMapLen 1024
-extern const int cloudBGMap[1024];
+extern const unsigned short cloudBGMap[1024];
 
 #endif // GRIT_LEVELBG_H
